dr in its own pkg
