+++ core/services/validators_test.go
@@ -1,17 +1,10 @@
 package services_test
 
 import (
-	"crypto/sha256"
 	"encoding/json"
 	"fmt"
 	"testing"
 	"time"
-
-	"github.com/smartcontractkit/chainlink/core/services/job"
-
-	"github.com/manyminds/api2go/jsonapi"
-
-	"github.com/smartcontractkit/chainlink/core/store/orm"
 
 	"github.com/smartcontractkit/chainlink/core/adapters"
 	"github.com/smartcontractkit/chainlink/core/assets"
@@ -488,348 +481,4 @@ func TestValidateInitiator_FeedsErrors(t *testing.T) {
 			require.Error(t, err)
 		})
 	}
-}
-
-func TestValidateOracleSpec(t *testing.T) {
-	var tt = []struct {
-		name       string
-		toml       string
-		setGlobals func(t *testing.T, c *orm.Config)
-		assertion  func(t *testing.T, os job.SpecDB, err error)
-	}{
-		{
-			name: "minimal non-bootstrap oracle spec",
-			toml: `
-type               = "offchainreporting"
-schemaVersion      = 1
-contractAddress    = "0x613a38AC1659769640aaE063C651F48E0250454C"
-isBootstrapPeer    = false
-observationSource = """
-ds1          [type=bridge name=voter_turnout];
-ds1_parse    [type=jsonparse path="one,two"];
-ds1_multiply [type=multiply times=1.23];
-ds1 -> ds1_parse -> ds1_multiply -> answer1;
-answer1      [type=median index=0];
-"""
-`,
-			assertion: func(t *testing.T, os job.SpecDB, err error) {
-				require.NoError(t, err)
-				// Should be able to jsonapi marshal/unmarshal the minimum spec.
-				// This ensures the UnmarshalJSON's defined on the fields handle a min spec correctly.
-				b, err := jsonapi.Marshal(os.OffchainreportingOracleSpec)
-				require.NoError(t, err)
-				var r job.OffchainReportingOracleSpec
-				err = jsonapi.Unmarshal(b, &r)
-				require.NoError(t, err)
-			},
-		},
-		{
-			name: "decodes valid oracle spec toml",
-			toml: `
-type               = "offchainreporting"
-schemaVersion      = 1
-contractAddress    = "0x613a38AC1659769640aaE063C651F48E0250454C"
-p2pPeerID          = "12D3KooWHfYFQ8hGttAYbMCevQVESEQhzJAqFZokMVtom8bNxwGq"
-p2pBootstrapPeers  = [
-"/dns4/chain.link/tcp/1234/p2p/16Uiu2HAm58SP7UL8zsnpeuwHfytLocaqgnyaYKP8wu7qRdrixLju",
-]
-isBootstrapPeer    = false
-keyBundleID        = "73e8966a78ca09bb912e9565cfb79fbe8a6048fab1f0cf49b18047c3895e0447"
-monitoringEndpoint = "chain.link:4321"
-transmitterAddress = "0xaA07d525B4006a2f927D79CA78a23A8ee680A32A"
-observationTimeout = "10s"
-observationSource = """
-ds1          [type=bridge name=voter_turnout];
-ds1_parse    [type=jsonparse path="one,two"];
-ds1_multiply [type=multiply times=1.23];
-ds1 -> ds1_parse -> ds1_multiply -> answer1;
-answer1      [type=median index=0];
-"""
-`,
-			assertion: func(t *testing.T, os job.SpecDB, err error) {
-				require.NoError(t, err)
-				assert.Equal(t, 1, int(os.SchemaVersion))
-				assert.False(t, os.OffchainreportingOracleSpec.IsBootstrapPeer)
-			},
-		},
-		{
-			name: "decodes bootstrap toml",
-			toml: `
-type               = "offchainreporting"
-schemaVersion      = 1
-contractAddress    = "0x613a38AC1659769640aaE063C651F48E0250454C"
-p2pPeerID          = "12D3KooWHfYFQ8hGttAYbMCevQVESEQhzJAqFZokMVtom8bNxwGq"
-p2pBootstrapPeers  = []
-isBootstrapPeer    = true
-`,
-			assertion: func(t *testing.T, os job.SpecDB, err error) {
-				require.NoError(t, err)
-				assert.Equal(t, 1, int(os.SchemaVersion))
-				assert.True(t, os.OffchainreportingOracleSpec.IsBootstrapPeer)
-			},
-		},
-		{
-			name: "raises error on extra keys",
-			toml: `
-type               = "offchainreporting"
-schemaVersion      = 1
-contractAddress    = "0x613a38AC1659769640aaE063C651F48E0250454C"
-p2pPeerID          = "12D3KooWHfYFQ8hGttAYbMCevQVESEQhzJAqFZokMVtom8bNxwGq"
-p2pBootstrapPeers  = [
-"/dns4/chain.link/tcp/1234/p2p/16Uiu2HAm58SP7UL8zsnpeuwHfytLocaqgnyaYKP8wu7qRdrixLju",
-]
-isBootstrapPeer    = true
-keyBundleID        = "73e8966a78ca09bb912e9565cfb79fbe8a6048fab1f0cf49b18047c3895e0447"
-monitoringEndpoint = "chain.link:4321"
-transmitterAddress = "0xaA07d525B4006a2f927D79CA78a23A8ee680A32A"
-observationTimeout = "10s"
-observationSource = """
-ds1          [type=bridge name=voter_turnout];
-ds1_parse    [type=jsonparse path="one,two"];
-ds1_multiply [type=multiply times=1.23];
-ds1 -> ds1_parse -> ds1_multiply -> answer1;
-answer1      [type=median index=0];
-"""
-`,
-			assertion: func(t *testing.T, os job.SpecDB, err error) {
-				require.Error(t, err)
-				assert.Contains(t, err.Error(), "unrecognised key for bootstrap peer: observationSource")
-			},
-		},
-		{
-			name: "empty pipeline string non-bootstrap node",
-			toml: `
-type               = "offchainreporting"
-schemaVersion      = 1
-contractAddress    = "0x613a38AC1659769640aaE063C651F48E0250454C"
-p2pPeerID          = "12D3KooWHfYFQ8hGttAYbMCevQVESEQhzJAqFZokMVtom8bNxwGq"
-p2pBootstrapPeers  = []
-isBootstrapPeer    = false
-`,
-			assertion: func(t *testing.T, os job.SpecDB, err error) {
-				require.Error(t, err)
-			},
-		},
-		{
-			name: "invalid dot",
-			toml: `
-type               = "offchainreporting"
-schemaVersion      = 1
-contractAddress    = "0x613a38AC1659769640aaE063C651F48E0250454C"
-p2pPeerID          = "12D3KooWHfYFQ8hGttAYbMCevQVESEQhzJAqFZokMVtom8bNxwGq"
-p2pBootstrapPeers  = []
-isBootstrapPeer    = false
-observationSource = """
-->
-"""
-`,
-			assertion: func(t *testing.T, os job.SpecDB, err error) {
-				require.Error(t, err)
-			},
-		},
-		{
-			name: "invalid peer address",
-			toml: `
-type               = "offchainreporting"
-schemaVersion      = 1
-contractAddress    = "0x613a38AC1659769640aaE063C651F48E0250454C"
-p2pPeerID          = "12D3KooWHfYFQ8hGttAYbMCevQVESEQhzJAqFZokMVtom8bNxwGq"
-p2pBootstrapPeers  = ["/invalid/peer/address"]
-isBootstrapPeer    = false
-observationSource = """
-blah
-"""
-`,
-			assertion: func(t *testing.T, os job.SpecDB, err error) {
-				require.Error(t, err)
-			},
-		},
-		{
-			name: "non-zero timeouts",
-			toml: `
-type               = "offchainreporting"
-schemaVersion      = 1
-contractAddress    = "0x613a38AC1659769640aaE063C651F48E0250454C"
-p2pPeerID          = "12D3KooWHfYFQ8hGttAYbMCevQVESEQhzJAqFZokMVtom8bNxwGq"
-p2pBootstrapPeers  = ["/dns4/chain.link/tcp/1234/p2p/16Uiu2HAm58SP7UL8zsnpeuwHfytLocaqgnyaYKP8wu7qRdrixLju"]
-isBootstrapPeer    = false
-blockchainTimeout  = "0s"
-observationSource = """
-blah
-"""
-`,
-			assertion: func(t *testing.T, os job.SpecDB, err error) {
-				require.Error(t, err)
-			},
-		},
-		{
-			name: "non-zero intervals",
-			toml: `
-type               = "offchainreporting"
-schemaVersion      = 1
-contractAddress    = "0x613a38AC1659769640aaE063C651F48E0250454C"
-p2pPeerID          = "12D3KooWHfYFQ8hGttAYbMCevQVESEQhzJAqFZokMVtom8bNxwGq"
-p2pBootstrapPeers  = ["/dns4/chain.link/tcp/1234/p2p/16Uiu2HAm58SP7UL8zsnpeuwHfytLocaqgnyaYKP8wu7qRdrixLju"]
-isBootstrapPeer    = false
-contractConfigTrackerSubscribeInterval = "0s"
-observationSource = """
-blah
-"""
-`,
-			assertion: func(t *testing.T, os job.SpecDB, err error) {
-				require.Error(t, err)
-			},
-		},
-		{
-			name: "broken monitoring endpoint",
-			toml: `
-type               = "offchainreporting"
-schemaVersion      = 1
-contractAddress    = "0x613a38AC1659769640aaE063C651F48E0250454C"
-p2pPeerID          = "12D3KooWHfYFQ8hGttAYbMCevQVESEQhzJAqFZokMVtom8bNxwGq"
-p2pBootstrapPeers  = []
-isBootstrapPeer    = true
-monitoringEndpoint = "\t/fd\2ff )(*&^%$#@"
-`,
-			assertion: func(t *testing.T, os job.SpecDB, err error) {
-				require.EqualError(t, err, "(8, 23): invalid escape sequence: \\2")
-			},
-		},
-		{
-			name: "max task duration > observation timeout should error",
-			toml: `
-type               = "offchainreporting"
-maxTaskDuration    = "30s"
-schemaVersion      = 1
-contractAddress    = "0x613a38AC1659769640aaE063C651F48E0250454C"
-p2pPeerID          = "12D3KooWHfYFQ8hGttAYbMCevQVESEQhzJAqFZokMVtom8bNxwGq"
-p2pBootstrapPeers  = [
-"/dns4/chain.link/tcp/1234/p2p/16Uiu2HAm58SP7UL8zsnpeuwHfytLocaqgnyaYKP8wu7qRdrixLju",
-]
-isBootstrapPeer    = false
-keyBundleID        = "73e8966a78ca09bb912e9565cfb79fbe8a6048fab1f0cf49b18047c3895e0447"
-monitoringEndpoint = "chain.link:4321"
-transmitterAddress = "0xaA07d525B4006a2f927D79CA78a23A8ee680A32A"
-observationTimeout = "10s"
-observationSource = """
-ds1          [type=bridge name=voter_turnout];
-"""
-`,
-			assertion: func(t *testing.T, os job.SpecDB, err error) {
-				require.Error(t, err)
-				require.Contains(t, err.Error(), "max task duration must be < observation timeout")
-			},
-		},
-		{
-			name: "invalid peer ID",
-			toml: `
-type               = "offchainreporting"
-schemaVersion      = 1
-contractAddress    = "0x613a38AC1659769640aaE063C651F48E0250454C"
-p2pPeerID = "blah"
-isBootstrapPeer    = true
-`,
-			assertion: func(t *testing.T, os job.SpecDB, err error) {
-				require.Error(t, err)
-				require.Contains(t, err.Error(), "failed to parse peer ID")
-			},
-		},
-		{
-			name: "individual max task duration > observation timeout should error",
-			toml: `
-type               = "offchainreporting"
-schemaVersion      = 1
-contractAddress    = "0x613a38AC1659769640aaE063C651F48E0250454C"
-p2pPeerID          = "12D3KooWHfYFQ8hGttAYbMCevQVESEQhzJAqFZokMVtom8bNxwGq"
-p2pBootstrapPeers  = [
-"/dns4/chain.link/tcp/1234/p2p/16Uiu2HAm58SP7UL8zsnpeuwHfytLocaqgnyaYKP8wu7qRdrixLju",
-]
-isBootstrapPeer    = false
-keyBundleID        = "73e8966a78ca09bb912e9565cfb79fbe8a6048fab1f0cf49b18047c3895e0447"
-monitoringEndpoint = "chain.link:4321"
-transmitterAddress = "0xaA07d525B4006a2f927D79CA78a23A8ee680A32A"
-observationTimeout = "10s"
-observationSource = """
-ds1          [type=bridge name=voter_turnout timeout="30s"];
-"""
-`,
-			assertion: func(t *testing.T, os job.SpecDB, err error) {
-				require.Error(t, err)
-				require.Contains(t, err.Error(), "individual max task duration must be < observation timeout")
-			},
-		},
-		{
-			name: "toml parse doesn't panic",
-			toml: string(cltest.MustHexDecodeString("2222220d5c22223b22225c0d21222222")),
-			assertion: func(t *testing.T, os job.SpecDB, err error) {
-				require.Error(t, err)
-			},
-		},
-		{
-			name: "invalid global default",
-			toml: `
-type               = "offchainreporting"
-schemaVersion      = 1
-contractAddress    = "0x613a38AC1659769640aaE063C651F48E0250454C"
-p2pPeerID          = "12D3KooWHfYFQ8hGttAYbMCevQVESEQhzJAqFZokMVtom8bNxwGq"
-p2pBootstrapPeers  = [
-"/dns4/chain.link/tcp/1234/p2p/16Uiu2HAm58SP7UL8zsnpeuwHfytLocaqgnyaYKP8wu7qRdrixLju",
-]
-isBootstrapPeer    = false
-keyBundleID        = "73e8966a78ca09bb912e9565cfb79fbe8a6048fab1f0cf49b18047c3895e0447"
-monitoringEndpoint = "chain.link:4321"
-transmitterAddress = "0xaA07d525B4006a2f927D79CA78a23A8ee680A32A"
-observationSource = """
-ds1          [type=bridge name=voter_turnout];
-ds1_parse    [type=jsonparse path="one,two"];
-ds1_multiply [type=multiply times=1.23];
-ds1 -> ds1_parse -> ds1_multiply -> answer1;
-answer1      [type=median index=0];
-"""
-`,
-			assertion: func(t *testing.T, os job.SpecDB, err error) {
-				require.Error(t, err)
-			},
-			setGlobals: func(t *testing.T, c *orm.Config) {
-				c.Set("OCR_OBSERVATION_TIMEOUT", "20m")
-			},
-		},
-	}
-
-	for _, tc := range tt {
-		t.Run(tc.name, func(t *testing.T) {
-			c := orm.NewConfig()
-			if tc.setGlobals != nil {
-				tc.setGlobals(t, c)
-			}
-			s, err := services.ValidatedOracleSpecToml(c, tc.toml)
-			tc.assertion(t, s, err)
-		})
-	}
-}
-
-func TestValidatedDirectRequestSpec(t *testing.T) {
-	toml := `
-type                = "directrequest"
-schemaVersion       = 1
-name                = "example eth request event spec"
-contractAddress     = "0x613a38AC1659769640aaE063C651F48E0250454C"
-observationSource   = """
-    ds1          [type=http method=GET url="example.com" allowunrestrictednetworkaccess="true"];
-    ds1_parse    [type=jsonparse path="USD"];
-    ds1_multiply [type=multiply times=100];
-    ds1 -> ds1_parse -> ds1_multiply;
-"""
-`
-
-	s, err := services.ValidatedDirectRequestSpec(toml)
-	require.NoError(t, err)
-
-	sha := sha256.Sum256([]byte(toml))
-
-	require.Equal(t, int32(0), s.ID)
-	require.Equal(t, "0x613a38AC1659769640aaE063C651F48E0250454C", s.DirectRequestSpec.ContractAddress.Hex())
-	require.Equal(t, sha[:], s.DirectRequestSpec.OnChainJobSpecID[:])
-	require.Equal(t, time.Time{}, s.DirectRequestSpec.CreatedAt)
-	require.Equal(t, time.Time{}, s.DirectRequestSpec.UpdatedAt)
 }
